Se cmabio el telefono y campos de valor y telefono a tipo Long para que no genere conflictos, ademas se intento mejorar la interfaz.

+++ Horizontes_Urbanos/src/main/java/CONTROLADOR/GestionarAsesor.java
@@ -37,7 +37,7 @@ public class GestionarAsesor {
     }
     
     public void editar(Asesor asesor, Long cedula,String nombre,String direccion,String correo,
-            int telefono) throws Exception {
+            Long telefono) throws Exception {
         
         asesor.setCedula(cedula);
         asesor.setNombre(nombre);
