+++ Horizontes_Urbanos/src/main/java/VISTA/admin/administrarApartamento.java
@@ -192,7 +192,7 @@ public class administrarApartamento extends javax.swing.JFrame {
         btnGuardarApartamento.setEnabled(false);
 
         int numeroApartamento = Integer.parseInt(txtNumeroApartamento.getText());
-        int valorApartamento = Integer.parseInt(txtValorApartamento.getText());
+        Long valorApartamento = Long.parseLong(txtValorApartamento.getText());
         String tipoUnidad = txtTipoUnidad.getText();
         double area = Double.parseDouble(txtArea.getText());
         String matricula = txtMatricula.getText();
